requirements.txt
AppKit==0.2.8
appnope==0.1.4
argcomplete==3.6.3
argon2_cffi_bindings==25.1.0
astroid==4.1.1
asyncssh_unofficial==0.9.2
atheris==3.0.0
BeaEnginePython==3.1.0
black==26.1.0
bokeh==3.8.2
brotli==1.2.0
brotlicffi==1.2.0.0
cached_property==2.0.1
cairocffi==1.7.1
capstone==5.0.7
Carbon==1.1.10
cchardet==2.1.7
cerealizer==0.8.4
cffi==2.0.0
clr==1.0.3
ConfigParser==7.2.0
contextlib2==21.6.0
copier==9.12.0
coverage==7.13.4
ctypes_snappy==1.03
curio==1.6
Cython==3.2.4
defcon==0.12.2
distorm3==3.5.2
dl==0.1.0
docrepr==0.2.0
exceptiongroup==1.3.1
flaky==3.8.1
Foundation==0.1.0a0.dev1
freetype_py==2.5.1
fs==2.4.16
gevent==25.9.1
glyphsLib==6.12.7
gobject==0.1.0
greenlet==3.3.2
h2==4.3.0
HTMLParser==0.0.2
hypothesis==6.151.9
importlib_resources==6.5.2
inflect==7.5.0
ini2toml==0.15
interegular==0.3.3
ipyparallel==9.0.2
jnius==1.1.0
js2py==0.74
jsonpath_ng==1.8.0
jupyter_ai==2.31.7
jupyter_ai_magics==2.31.7
jupyter_collaboration==4.2.1
jupyterhub==5.4.3
keyring==25.7.0
lz4==4.4.5
mtrand==0.1
munkres==1.1.4
mypy==1.19.1
netifaces==0.11.0
nose==1.3.7
numpydoc==1.10.0
olefile==0.47
openapi_core==0.22.0
outcome==1.3.0.post0
pandas==3.0.1
pexpect==4.9.0
playwright==1.58.0
psyco==1.6
pycares==5.0.1
pycurl==7.45.7
pyczmq==0.0.4
pydot==4.0.1
pygraphviz==1.14
PyInstaller==6.19.0
pyodide==0.0.2
PyOpenGL==3.1.10
pyperf==2.10.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.10.2
Pyside==1.2.4
PySide==1.2.4
PySide2==5.15.2.1
PySide6==6.10.2
pysqlite==2.8.3
pytest_run_parallel==0.8.2
python_socks==2.8.1
Quartz==0.0.1.dev0
railroad==0.5.0
regex==2026.2.28
reportlab==4.4.10
Res==0.1.7
rfc3987==1.3.8
ruamel.base==1.0.0
scipy==1.17.1
scipy_doctest==2.1.0
selenium==4.41.0
setuptools_scm==9.2.2
shiboken2==5.15.2.1
shiboken6==6.10.2
sip==6.15.1
sniffio==1.3.1
socksio==1.0.0
Sphinx==9.1.0
sphinx_gallery==0.20.0
SQLAlchemy==2.0.48
stackless==0.0.1
sympy==1.14.0
testpath==0.6.0
threadpoolctl==3.6.0
trove_classifiers==2026.1.14.14
ufoLib2==0.18.1
uharfbuzz==0.53.3
unicodedata2==17.0.1
urllib3_secure_extra==0.1.0
uvloop==0.22.1
websockets==16.0
winloop==0.5.0
wsaccel==0.6.7
xattr==1.3.0
xmlrpclib==1.0.1
xx==3.3.2
yapf==0.43.0
zopfli==0.4.1
zstandard==0.25.0
